grouped_cv_eb0/requirements.txt
# Core ML & Data Science
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Image Processing
Pillow>=8.3.0
opencv-python>=4.5.0
albumentations>=1.1.0

# Hyperparameter Tuning
optuna>=2.10.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
openpyxl>=3.0.0
tqdm>=4.62.0
pyyaml>=5.4.0

# Optional: For CAM visualization
tensorboard>=2.7.0
